LPS-65040 Rename package for consistency
