+++ modules/util/json-web-service-client/src/main/java/com/liferay/json/web/service/client/JSONWebServiceClient.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.jsonwebserviceclient;
+package com.liferay.json.web.service.client;
 
 import java.security.KeyStore;
