+++ modules/util/json-web-service-client/src/main/java/com/liferay/json/web/service/client/JSONWebServiceClientImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.jsonwebserviceclient;
+package com.liferay.json.web.service.client;
 
 import java.io.IOException;
 import java.io.InterruptedIOException;
